pyproject.toml
[tool.poetry]
name = "fake-t-shirt-spanner-cdc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Faker = "20.1.0"
google-cloud-spanner = "3.40.1"
google-cloud-pubsub = "2.18.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
